Add openEmu to list of English.
OpenEmu is about to change the world of video game emulation. One console at a time...

For the first time, the 'It just works' philosophy now extends to open
source video game emulation on the Mac. With OpenEmu, it is extremely
easy to add, browse, organize and with a compatible gamepad, play those
favorite games (ROMs) you already own.
http://openemu.org/
https://github.com/OpenEmu